+++ packages/expo-router/CHANGELOG.md
@@ -52,6 +52,7 @@
 
 ### 🎉 New features
 
+- Export `createBaseTabProps`, `createNativeStackProps`, `createJSStackProps`, `createJSTabsProps`, `createJSTopTabsProps`, and `createNativeTabsProps` for integrating navigators with Expo Router.
 - Export `createBaseStackProps` for integrating custom stack navigators with Expo Router.
 - Expose route removal prevention state to custom navigators through `createProps`.
 - Expose route preload state to custom navigators through `createProps`.